requirements.txt
wikipedia
pageviewapi.period
